// packages/webapp/src/interfaces.d.ts
/*!
 * Copyright (c) Microsoft. All rights reserved.
 * Licensed under the MIT license. See LICENSE file in the project.
 */
export declare type NodeId = string | number
export interface Node {
	id: NodeId
	weight: number
	community?: number
	x: number
	y: number
}
export interface Edge {
	source: NodeId
	target: NodeId
	weight: number
}
export interface Graph {
	nodes: Node[]
	edges: Edge[]
}